requirements.txt
fastapi
uvicorn[standard]
web3
gspread
python-dotenv
python-binance
cryptography
fernet